+++ pmd-core/src/main/java/net/sourceforge/pmd/Report.java
@@ -8,13 +8,9 @@ import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 
 import net.sourceforge.pmd.renderers.AbstractAccumulatingRenderer;
 
@@ -23,17 +19,14 @@ import net.sourceforge.pmd.renderers.AbstractAccumulatingRenderer;
  * includes violations, suppressed violations, metrics, error during processing
  * and configuration errors.
  */
-public class Report implements Iterable<RuleViolation> {
+public class Report {
 
-    // Note that this and the above data structure are both being maintained for
-    // a bit
-    private final List<RuleViolation> violations = new ArrayList<>();
     private final List<ThreadSafeReportListener> listeners = new ArrayList<>();
-    private List<ProcessingError> errors;
-    private List<ConfigurationError> configErrors;
-    private long start;
-    private long end;
-    private List<SuppressedViolation> suppressedRuleViolations = new ArrayList<>();
+
+    private final List<RuleViolation> violations = new ArrayList<>();
+    private final List<SuppressedViolation> suppressedRuleViolations = new ArrayList<>();
+    private final List<ProcessingError> errors = new ArrayList<>();
+    private final List<ConfigurationError> configErrors = new ArrayList<>();
 
     /**
      * Creates a new, initialized, empty report for the given file name.
@@ -53,51 +46,6 @@ public class Report implements Iterable<RuleViolation> {
         ctx.setReport(report);
         ctx.setSourceCodeFile(new File(fileName));
         return report;
-    }
-
-    /**
-     * Represents a duration. Useful for reporting processing time.
-     */
-    public static class ReadableDuration {
-        private final long duration;
-
-        /**
-         * Creates a new duration.
-         *
-         * @param duration
-         *            the duration in milliseconds.
-         */
-        public ReadableDuration(long duration) {
-            this.duration = duration;
-        }
-
-        /**
-         * Gets a human readable representation of the duration, such as "1h 3m
-         * 5s".
-         *
-         * @return human readable representation of the duration
-         */
-        public String getTime() {
-            assert this.duration >= 0;
-            Duration duration = Duration.ofMillis(this.duration);
-            long seconds = duration.getSeconds();
-
-            long hours = seconds / 3600;
-            seconds -= hours * 3600;
-            long minutes = seconds / 60;
-            seconds -= minutes * 60;
-
-            StringBuilder res = new StringBuilder();
-            if (hours > 0) {
-                res.append(hours).append("h ");
-            }
-            if (hours > 0 || minutes > 0) {
-                res.append(minutes).append("m ");
-            }
-            res.append(seconds).append('s');
-
-            return res.toString();
-        }
     }
 
     /**
@@ -183,43 +131,8 @@ public class Report implements Iterable<RuleViolation> {
         }
     }
 
-
-    /**
-     * Calculate a summary of violations per rule.
-     *
-     * @return a Map summarizing the Report: String (rule name) -&gt; Integer (count
-     *         of violations)
-     */
-    public Map<String, Integer> getSummary() {
-        Map<String, Integer> summary = new HashMap<>();
-        for (RuleViolation rv : violations) {
-            String name = rv.getRule().getName();
-            if (!summary.containsKey(name)) {
-                summary.put(name, 0);
-            }
-            Integer count = summary.get(name);
-            summary.put(name, count + 1);
-        }
-        return summary;
-    }
-
-    /**
-     * Registers a report listener
-     *
-     * @param listener
-     *            the listener
-     */
-    public void addListener(ThreadSafeReportListener listener) {
-        listeners.add(listener);
-    }
-
-    public List<SuppressedViolation> getSuppressedRuleViolations() {
-        return suppressedRuleViolations;
-    }
-
     /**
      * Represents a violation, that has been suppressed.
-     * TODO this should implement RuleViolation
      */
     public static class SuppressedViolation {
         private final RuleViolation rv;
@@ -252,16 +165,19 @@ public class Report implements Iterable<RuleViolation> {
         }
     }
 
-
-    public void addSuppressedViolation(SuppressedViolation sv) {
-        suppressedRuleViolations.add(sv);
+    /**
+     * Registers a report listener
+     *
+     * @param listener the listener
+     */
+    public void addListener(ThreadSafeReportListener listener) {
+        listeners.add(listener);
     }
 
     /**
      * Adds a new rule violation to the report and notify the listeners.
      *
-     * @param violation
-     *            the violation to add
+     * @param violation the violation to add
      */
     public void addRuleViolation(RuleViolation violation) {
         int index = Collections.binarySearch(violations, violation, RuleViolationComparator.INSTANCE);
@@ -272,15 +188,18 @@ public class Report implements Iterable<RuleViolation> {
     }
 
     /**
+     * Adds a new suppressed violation.
+     */
+    public void addSuppressedViolation(SuppressedViolation sv) {
+        suppressedRuleViolations.add(sv);
+    }
+
+    /**
      * Adds a new configuration error to the report.
      *
-     * @param error
-     *            the error to add
+     * @param error the error to add
      */
     public void addConfigError(ConfigurationError error) {
-        if (configErrors == null) {
-            configErrors = new ArrayList<>();
-        }
         configErrors.add(error);
     }
 
@@ -291,9 +210,6 @@ public class Report implements Iterable<RuleViolation> {
      *            the error to add
      */
     public void addError(ProcessingError error) {
-        if (errors == null) {
-            errors = new ArrayList<>();
-        }
         errors.add(error);
     }
 
@@ -307,99 +223,50 @@ public class Report implements Iterable<RuleViolation> {
      * @see AbstractAccumulatingRenderer
      */
     public void merge(Report r) {
-        Iterator<ProcessingError> i = r.errors();
-        while (i.hasNext()) {
-            addError(i.next());
-        }
-        Iterator<ConfigurationError> ce = r.configErrors();
-        while (ce.hasNext()) {
-            addConfigError(ce.next());
-        }
-        for (RuleViolation violation : r) {
+        errors.addAll(r.errors);
+        configErrors.addAll(r.configErrors);
+        suppressedRuleViolations.addAll(r.suppressedRuleViolations);
+
+        for (RuleViolation violation : r.getViolations()) {
             int index = Collections.binarySearch(violations, violation, RuleViolationComparator.INSTANCE);
             violations.add(index < 0 ? -index - 1 : index, violation);
         }
-        suppressedRuleViolations.addAll(r.getSuppressedRuleViolations());
     }
 
-    public boolean isEmpty() {
-        return !violations.iterator().hasNext() && !hasErrors();
+
+    /**
+     * Returns an unmodifiable list of violations that were suppressed.
+     */
+    public List<SuppressedViolation> getSuppressedRuleViolations() {
+        return Collections.unmodifiableList(suppressedRuleViolations);
     }
 
     /**
-     * Checks whether any processing errors have been reported.
-     *
-     * @return <code>true</code> if there were any processing errors,
-     *         <code>false</code> otherwise
+     * Returns an unmodifiable list of violations that have been
+     * recorded until now. None of those violations were suppressed.
      */
-    public boolean hasErrors() {
-        return errors != null && !errors.isEmpty();
+    public List<RuleViolation> getViolations() {
+        return Collections.unmodifiableList(violations);
     }
+
 
     /**
-     * Checks whether any configuration errors have been reported.
-     *
-     * @return <code>true</code> if there were any configuration errors,
-     *         <code>false</code> otherwise
+     * Returns an unmodifiable list of processing errors that have been
+     * recorded until now.
      */
-    public boolean hasConfigErrors() {
-        return configErrors != null && !configErrors.isEmpty();
+    public List<ProcessingError> getProcessingErrors() {
+        return Collections.unmodifiableList(errors);
     }
 
-    @Override
-    public Iterator<RuleViolation> iterator() {
-        return violations.iterator();
-    }
 
     /**
-     * Returns an iterator of the reported processing errors.
-     *
-     * @return the iterator
+     * Returns an unmodifiable list of configuration errors that have
+     * been recorded until now.
      */
-    public Iterator<ProcessingError> errors() {
-        return errors == null ? Collections.<ProcessingError>emptyIterator() : errors.iterator();
+    public List<ConfigurationError> getConfigErrors() {
+        return Collections.unmodifiableList(configErrors);
     }
 
-    /**
-     * Returns an iterator of the reported configuration errors.
-     *
-     * @return the iterator
-     */
-    public Iterator<ConfigurationError> configErrors() {
-        return configErrors == null ? Collections.<ConfigurationError>emptyIterator() : configErrors.iterator();
-    }
-
-    /**
-     * The number of violations.
-     *
-     * @return number of violations.
-     */
-    public int size() {
-        return violations.size();
-    }
-
-    /**
-     * Mark the start time of the report. This is used to get the elapsed time
-     * in the end.
-     *
-     * @see #getElapsedTimeInMillis()
-     */
-    public void start() {
-        start = System.currentTimeMillis();
-    }
-
-    /**
-     * Mark the end time of the report. This is ued to get the elapsed time.
-     *
-     * @see #getElapsedTimeInMillis()
-     */
-    public void end() {
-        end = System.currentTimeMillis();
-    }
-
-    public long getElapsedTimeInMillis() {
-        return end - start;
-    }
 
     public List<ThreadSafeReportListener> getListeners() {
         return listeners;