+++ pmd-java/src/test/java/net/sourceforge/pmd/ExcludeLinesTest.java
@@ -43,7 +43,7 @@ public class ExcludeLinesTest extends RuleTst {
         ctx.setLanguageVersion(LanguageRegistry.getLanguage(JavaLanguageModule.NAME).getDefaultVersion());
         RuleSet rules = RulesetsFactoryUtils.defaultFactory().createSingleRuleRuleSet(rule);
         p.getSourceCodeProcessor().processSourceCode(new StringReader(TEST3), new RuleSets(rules), ctx);
-        assertTrue(r.isEmpty());
+        assertTrue(r.getViolations().isEmpty());
         assertEquals(r.getSuppressedRuleViolations().size(), 1);
     }