+++ pmd-java/src/test/java/net/sourceforge/pmd/ReportTest.java
@@ -23,7 +23,7 @@ public class ReportTest extends RuleTst {
     public void testBasic() {
         Report r = new Report();
         runTestFromString(TEST1, new FooRule(), r, defaultLanguage);
-        assertFalse(r.isEmpty());
+        assertFalse(r.getViolations().isEmpty());
     }
 
     @Test
@@ -32,7 +32,7 @@ public class ReportTest extends RuleTst {
         Rule rule = new FooRule();
         rule.setProperty(Rule.VIOLATION_SUPPRESS_REGEX_DESCRIPTOR, ".*No Foo.*");
         runTestFromString(TEST1, rule, rpt, defaultLanguage);
-        assertTrue(rpt.isEmpty());
+        assertTrue(rpt.getViolations().isEmpty());
         assertEquals(1, rpt.getSuppressedRuleViolations().size());
     }
 
@@ -42,7 +42,7 @@ public class ReportTest extends RuleTst {
         Rule rule = new FooRule();
         rule.setProperty(Rule.VIOLATION_SUPPRESS_XPATH_DESCRIPTOR, ".[@SimpleName = 'Foo']");
         runTestFromString(TEST1, rule, rpt, defaultLanguage);
-        assertTrue(rpt.isEmpty());
+        assertTrue(rpt.getViolations().isEmpty());
         assertEquals(1, rpt.getSuppressedRuleViolations().size());
     }
 
@@ -51,7 +51,7 @@ public class ReportTest extends RuleTst {
         Report rpt = new Report();
         runTestFromString(TEST2, new FooRule(), rpt,
                 LanguageRegistry.getLanguage(JavaLanguageModule.NAME).getVersion("1.5"));
-        assertTrue(rpt.isEmpty());
+        assertTrue(rpt.getViolations().isEmpty());
         assertEquals(1, rpt.getSuppressedRuleViolations().size());
     }
 
@@ -60,7 +60,7 @@ public class ReportTest extends RuleTst {
         Report rpt = new Report();
         runTestFromString(TEST2_FULL, new FooRule(), rpt,
                 LanguageRegistry.getLanguage(JavaLanguageModule.NAME).getVersion("1.5"));
-        assertTrue(rpt.isEmpty());
+        assertTrue(rpt.getViolations().isEmpty());
         assertEquals(1, rpt.getSuppressedRuleViolations().size());
     }
 
@@ -68,7 +68,7 @@ public class ReportTest extends RuleTst {
     public void testExclusionsInReportWithNOPMD() {
         Report rpt = new Report();
         runTestFromString(TEST3, new FooRule(), rpt, defaultLanguage);
-        assertTrue(rpt.isEmpty());
+        assertTrue(rpt.getViolations().isEmpty());
         assertEquals(1, rpt.getSuppressedRuleViolations().size());
     }